Use "public-api" for URI instead of version no
Brightpearl are deprecating the use of specific version numbers in the URI for their public API. Instead, you should use "public-api" in the same location within the URI.

- Change "2.0.0" to "public-api".
- Don't allow this change to be overridden in configuration.

This change will result in the config ignoring API versions being set, so the change is non-breaking.

I have tested the new URI's and they work.